feat: add recoverable secrets.
logion-network/logion-internal#1251
