+++ resources/mail/recoverable-secret-added.pug
@@ -1,0 +1,12 @@
+| logion notification - Recoverable secret added
+| Dear #{walletUser.firstName} #{walletUser.lastName},
+|
+| You receive this message because you just added a recoverable secret with name #{secretName} to your Identity LOC #{loc.id}.
+|
+| Please note, you can't answer this email. The logion network will never ask you to provide any kind of information or access to a web address through its email notifications.
+|
+| As a reminder, find below all details of the Legal Officer you'll have to contact in order to recover the secret's value:
+|
+include /legal-officer-details.pug
+|
+include /footer.pug
